# Optimize agent timestamp deletion stage in Authd