|Related issue|
|---|
|Closes #9293|

This PR aims to optimize function `OS_RemoveAgentTimestamp`, which deletes the line with a given ID from file `queue/agents-timestamp`.

## Change

The old behavior read `agents-timestamp` line by line, and appended them into a buffer if the ID was different from the target agent ID. Then, the function dumped the buffer back into a temporary file and renamed it to `agents-timestamp`.

The new behavior reads `agents-timestamp` and dumps each line directly into the temporary file except if the ID matched the target agent ID.

### Measurements

|Function|Previous execution time|Current execution time|Improvement|
|---|---:|---:|---:|
|`OS_RemoveAgentTimestamp()`|7060 ms|**7439 µs**|99.89%|

## Tests

- [X] Build manager on Linux.
- [X] Verify that agents-timestamp remains correct.
- [X] Stress test for affected components.
- [ ] Run Authd on Valgrind.